requirements.txt
altair
annotated-types
anyio
attrs
backoff
bcrypt
blinker
Brotli
build
cachetools
certifi
charset-normalizer
chromadb
click
colorama
coloredlogs
distro
docopt
durationpy
fastapi
filelock
flatbuffers
fsspec
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
h11
httpcore
httplib2
httptools
httpx
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
Jinja2
joblib
jsonschema
jsonschema-specifications
kubernetes
markdown-it-py
MarkupSafe
mdurl
mmh3
mpmath
mutagen
narwhals
networkx
numpy
oauthlib
onnxruntime
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
overrides
packaging
pandas
pillow
pipreqs
posthog
proto-plus
protobuf
pyarrow
pyasn1
pyasn1_modules
pybase64
pycryptodomex
pydantic
pydantic_core
pydeck
Pygments
pyparsing
PyPika
pyproject_hooks
pyreadline3
PySide6
PySide6_Addons
PySide6_Essentials
python-dateutil
python-dotenv
pytube
pytz
PyYAML
referencing
regex
requests
requests-oauthlib
rich
rpds-py
rsa
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
shellingham
shiboken6
six
smmap
sniffio
starlette
streamlit
sympy
tenacity
threadpoolctl
tiktoken
tokenizers
toml
torch
tornado
tqdm
transformers
typer
typing-inspection
typing_extensions
tzdata
uritemplate
urllib3
uvicorn
watchdog
watchfiles
websocket-client
websockets
yarg
yt-dlp
zipp